+++ DataModels/VM/AircraftEquipment/AircraftEquipmentDatatableParams.cs
@@ -4,6 +4,6 @@ namespace DataModels.VM.AircraftEquipment
 {
     public class AircraftEquipmentDatatableParams : DatatableParams
     {
-        public int AircraftId { get; set; }
+        public long AircraftId { get; set; }
     }
 }
